hack-university-orm-master/requirements.txt
Django<1.10
psycopg2==2.6.1
ipython==4.1.2
django-extensions==1.6.1
django-annoying==0.9.0
